dev-requirements.txt
pip~=21.0.1
pylint~=2.7.0
flake8~=3.8.4
bandit~=1.7.0
mypy~=0.812
wheel~=0.36.2
tox~=3.22.0
coverage~=5.4
pytest-cov~=2.11.1
